+++ cmd/xray-entry/exec_linux.go
@@ -8,14 +8,26 @@ import (
 	"syscall"
 )
 
-func execXray(args []string) error {
-	return syscall.Exec(defaultXray, args, xrayEnvironment())
-}
-
 func execXrayWithStdin(f *os.File, args []string) error {
-	if int(f.Fd()) != 0 {
-		if err := syscall.Dup3(int(f.Fd()), 0, 0); err != nil {
-			return errors.New("connect generated configuration to standard input")
+	sourceFD := int(f.Fd())
+	duplicate := -1
+	if sourceFD == 0 {
+		var err error
+		duplicate, err = syscall.Dup(sourceFD)
+		if err != nil {
+			return errors.New("connect configuration to standard input")
+		}
+		sourceFD = duplicate
+	}
+	if err := syscall.Dup3(sourceFD, 0, 0); err != nil {
+		if duplicate >= 0 {
+			_ = syscall.Close(duplicate)
+		}
+		return errors.New("connect configuration to standard input")
+	}
+	if duplicate >= 0 {
+		if err := syscall.Close(duplicate); err != nil {
+			return errors.New("close duplicated configuration descriptor")
 		}
 	}
 	return syscall.Exec(defaultXray, args, xrayEnvironment())