+++ cmd/xray-entry/exec_unsupported.go
@@ -7,10 +7,6 @@ import (
 	"os"
 )
 
-func execXray(args []string) error {
-	return errors.New("xray-entry runtime is supported only on linux")
-}
-
 func execXrayWithStdin(f *os.File, args []string) error {
 	return errors.New("xray-entry runtime is supported only on linux")
 }